// apps/api/src/modules/communication/domain/provider-capability.ts
import { Channel } from './channel.enum';

export interface ProviderCapability {
  readonly supportedChannels: Channel[];
  readonly supportsHtml: boolean;
  readonly supportsAttachments: boolean;
  readonly supportsRichMedia: boolean;
  readonly supportsNativeTemplates: boolean;
  readonly maxPayloadBytes?: number;
}